pyproject.toml
[project]
name = "nonebot-plugin-otto-hzys"
version = "0.1.0"
description = "插件描述"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Misty02600", email = "Misty02600@gmail.com" }]
dependencies = [
  "httpx>=0.27.0,<1.0.0",
  "nonebot-plugin-alconna>=0.60.0,<1.0.0",    # 跨平台命令解析拓展
  # "nonebot-plugin-apscheduler>=0.5.0,<1.0.0", # 定时任务
  # "nonebot-plugin-localstore>=0.7.4,<1.0.0",  # 存储文件
  # "nonebot-plugin-uninfo>=0.10.0,<1.0.0",     # 多平台用户信息获取
  # "nonebot-adapter-onebot>=2.4.6,<3.0.0",
  "nonebot2>=2.4.2,<3.0.0",

]

[project.urls]
Homepage = "https://github.com/Misty02600/nonebot-plugin-otto-hzys"
Issues = "https://github.com/Misty02600/nonebot-plugin-otto-hzys/issues"
Repository = "https://github.com/Misty02600/nonebot-plugin-otto-hzys.git"

[dependency-groups]
dev = [
  "commitizen>=4.1.0",
  "git-cliff>=2.11.0,<3.0.0",
  "prek>=0.2.0",
  { include-group = "test" },
]
test = [
  "basedpyright>=1.16.0",
  "ruff>=0.14.13,<1.0.0",
  "nonebot-adapter-onebot>=2.4.6,<3.0.0",
  "nonebot2[fastapi]>=2.4.4,<3.0.0",
  "nonebug>=0.4.3,<1.0.0",
  "pytest-asyncio>=1.3.0,<1.4.0",
  "pytest-cov>=7.0.0",
]

[tool.nonebot.plugins]
nonebot-plugin-otto-hzys = ["nonebot_plugin_otto_hzys"]

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
  { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"                           # 当前版本
tag_format = "v$version"
version_files = ["pyproject.toml:^version"]
major_version_zero = true                   # 允许 0.x.x 版本号

[tool.coverage.report]
exclude_lines = [
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
  "@overload",
  "except ImportError:",
]

[tool.pyright]
include = ["src", "tests"] # 要检查的目录
pythonVersion = "3.11"     # 目标 Python 版本（影响可用的类型特性）
pythonPlatform = "All"     # 目标平台：All | Linux | Windows | Darwin
# typeCheckingMode 类型检查严格程度，可选值：
#   - "off"：      关闭类型检查
#   - "basic"：    基础检查（仅报告明确的类型错误）
#   - "standard"： 标准检查（推荐，平衡严格性与实用性）
#   - "strict"：   严格检查（要求完整的类型注解，适合新项目）
#   - "all"：      最严格（启用所有检查，可能产生大量警告）
typeCheckingMode = "standard"

# 测试目录的特殊配置（放宽部分规则以适应测试场景）
[[tool.pyright.executionEnvironments]]
root = "tests"
reportPrivateUsage = "none"      # 允许测试访问私有属性 (_variable)
reportUnknownMemberType = "none" # 允许 Mock 对象等动态类型

[tool.pytest.ini_options]
addopts = [
  "--import-mode=importlib", # 使用 importlib 导入模式（推荐，避免 sys.path 污染）
  "--strict-markers",        # 严格标记模式（未注册的 marker 会报错）
  "--tb=short",              # 简短的错误回溯（减少输出噪音）
  "-ra",                     # 显示所有非通过测试的摘要（a=all except passed）
]
testpaths = ["tests"] # 仅运行 tests 目录下的测试
pythonpath = ["src", "tests"] # 添加到 Python 路径，确保模块可导入
asyncio_mode = "auto" # 自动检测异步测试函数，无需手动标记 @pytest.mark.asyncio
asyncio_default_fixture_loop_scope = "session" # 每个测试函数使用独立的事件循环

[tool.ruff]
line-length = 88                 # 每行最大字符数（与 Black 保持一致）
src = ["src", "tests"]           # 源代码目录（用于判断导入是本地模块还是第三方模块）
include = ["src/**", "tests/**"] # 仅检查 src 和 tests 目录
extend-exclude = ["scripts"]     # 脚本目录不做 lint 检查

[tool.ruff.format]
line-ending = "lf" # 行尾符：lf（Unix）| crlf（Windows）| auto | native

[tool.ruff.lint]
# 启用的规则集，每个字母代码对应一组规则
select = [
  "F",     # Pyflakes：基础语法和逻辑错误检查
  "W",     # pycodestyle warnings：PEP 8 风格警告
  "E",     # pycodestyle errors：PEP 8 风格错误
  "I",     # isort：导入语句排序
  "B",     # flake8-bugbear：发现潜在 bug 和设计问题
  "UP",    # pyupgrade：升级到新版 Python 语法
  "ASYNC", # flake8-async：异步代码最佳实践
  "C4",    # flake8-comprehensions：推导式优化建议
  "T10",   # flake8-debugger：检测遗留的调试器语句（如 pdb）
  "T20",   # flake8-print：检测 print 语句（生产代码应使用 logging）
  "PYI",   # flake8-pyi：类型存根文件（.pyi）规范检查
  "PT",    # flake8-pytest-style：pytest 代码风格检查
  "Q",     # flake8-quotes：引号使用规范
  "TID",   # flake8-tidy-imports：导入语句整理
  "RUF",   # Ruff-specific：Ruff 特有的规则
]
# 忽略的规则（根据项目需求放宽）
ignore = [
  "E501",   # 行长度由 formatter 控制，lint 不重复检查
  "E402",   # 允许模块导入不在文件顶部（某些情况下需要先执行代码）
  "UP037",  # 允许使用引号包裹的类型注解（前向引用）
  "RUF001", # 允许字符串中的中文等 Unicode 字符
  "RUF002", # 允许文档字符串中的中文等 Unicode 字符
  "RUF003", # 允许注释中的中文等 Unicode 字符
  "W191",   # 允许使用制表符缩进
  "TID252", # 允许使用相对导入
  "B008",   # 允许在函数参数默认值中使用函数调用
]

[tool.ruff.lint.isort]
extra-standard-library = ["typing_extensions"] # 将 typing_extensions 视为标准库
